id改long，新建course

+++ SpringBootWork/src/main/java/cn/edu/zucc/controller/ClassroomController.java
@@ -28,7 +28,7 @@ public class ClassroomController {
         return classroomService.addClassroom(classroom);
     }
     @DeleteMapping("/{id}")
-    public Classroom deleteById(@PathVariable int id) {
+    public Classroom deleteById(@PathVariable Long id) {
         return classroomService.deleteById(id);
     }
