+++ SpringBootWork/src/main/java/cn/edu/zucc/entity/Teacher.java
@@ -10,8 +10,17 @@ import java.util.*;
 @Table(name = "teacher")
 public class Teacher{
     @Id
-    private int TeacherID;
+    private Long TeacherID;
+
     @ManyToOne
     @JoinColumn(name = "UserID")
     private User User;
+
+//    @ManyToMany
+//    @JoinTable(
+//            name = "teacher_select_course",
+//            joinColumns = @JoinColumn(name = "Teacherid"),
+//            inverseJoinColumns = @JoinColumn(name = "Courseid")
+//    )
+//    private Set<Course> coursesTaught = new HashSet<>();
 }
