+++ SpringBootWork/src/main/java/cn/edu/zucc/repository/ClassroomRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ClassroomRepository extends JpaRepository<Classroom, Integer> {
+public interface ClassroomRepository extends JpaRepository<Classroom, Long> {
     List<Classroom> findByClassroomName(String name);
 }
